renaming the controller
